*** empty log message ***
git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@396 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/util/ValueConverter.java
@@ -100,7 +100,7 @@ public class ValueConverter
 		{
 			case Types.BIGINT:
 				if (aValue.toString().length() == 0) return null;
-				return new BigInteger(aValue.toString().trim());
+				return new Long(aValue.toString().trim());
 			case Types.INTEGER:
 			case Types.SMALLINT:
 			case Types.TINYINT:
